Moves detail::TypeInfo to detail/TypeInfo.hpp
